Moved extension definitions to their own file.
